requirements-win.txt
#git+https://github.com/open-homeautomation/miflora.git#egg=miflora
miflora==0.7.2
bleak==0.19.5
btlewrap==0.0.10
paho-mqtt==1.4.0
wheel==0.29.0
sdnotify==0.3.1
colorama==0.3.9
Unidecode==0.4.21
